mediamanagementproject/requirements.txt
django>=4
djangorestframework==3.13.1
djangorestframework-xml
defusedxml
pyyaml
uritemplate
drf-spectacular
celery==5.2.7
flower==1.2.0
django-celery-results==2.4.0
pexpect
paramiko
wheel
pysam
django-debug-toolbar==3.6.0

